backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pillow==10.4.0
transformers==4.46.2
torch==2.5.1
python-multipart==0.0.12
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0